Rename builtins to logic. Add well_founded trait to it
