+++ creusot-contracts-proc/src/pretyping.rs
@@ -100,7 +100,7 @@ pub fn encode_term(term: RT) -> Result<TokenStream, EncodeError> {
         RT::Model(TermModel { term, .. }) => {
             let term = encode_term(*term)?;
             Ok(quote! {
-                creusot_contracts::builtins::Model::model(#term)
+                creusot_contracts::logic::Model::model(#term)
             })
         }
         RT::Verbatim(_) => todo!(),
