+++ creusot-contracts/src/logic.rs
@@ -2,7 +2,10 @@ mod int;
 mod model;
 mod resolve;
 mod seq;
+pub mod well_founded;
+
 pub use int::*;
 pub use model::*;
 pub use resolve::*;
 pub use seq::*;
+pub use well_founded::*;
