+++ creusot-contracts/src/logic/model.rs
