+++ creusot-contracts/src/logic/resolve.rs
