+++ creusot-contracts/src/std.rs
@@ -1,7 +1,5 @@
 pub mod clone;
 pub mod vec;
-pub mod well_founded;
 
 pub use clone::*;
 pub use vec::*;
-pub use well_founded::*;
