+++ creusot/tests/should_succeed/all_zero.stdout
@@ -75,30 +75,30 @@ module AllZero_Get
       | Type.AllZero_List_Nil -> Type.Core_Option_Option_None
       end
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module AllZero_AllZero_Interface
@@ -123,10 +123,10 @@ module AllZero_AllZero
   clone AllZero_Get as Get0
   clone AllZero_Len as Len0
   use mach.int.Int64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve2 with type t = Type.allzero_list
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = Type.allzero_list
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve2 with type t = Type.allzero_list
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = Type.allzero_list
   let rec cfg all_zero (l : borrowed (Type.allzero_list)) : ()
     ensures { Len0.len ( * l) = Len0.len ( ^ l) }
     ensures { forall i : (int) . 0 <= i && i < Len0.len ( * l) -> Get0.get ( ^ l) i = Type.Core_Option_Option_Some (0 : uint32) }
