+++ creusot/tests/should_succeed/binary_search.stdout
@@ -81,11 +81,11 @@ module BinarySearch_Impl0_Get
       | Type.BinarySearch_List_Nil -> Type.Core_Option_Option_None
       end
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -111,11 +111,11 @@ module BinarySearch_Impl0_Index
   clone BinarySearch_Impl0_Get as Get0 with type t = t
   clone BinarySearch_Impl0_LenLogic as LenLogic0 with type t = t
   use mach.int.Int64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve4 with type self = Type.binarysearch_list t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = Type.binarysearch_list t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve4 with type self = Type.binarysearch_list t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.binarysearch_list t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   let rec cfg index (self : Type.binarysearch_list t) (ix : usize) : t
     requires {UInt64.to_int ix < LenLogic0.len_logic self}
     ensures { Type.Core_Option_Option_Some result = Get0.get self (UInt64.to_int ix) }
@@ -232,10 +232,10 @@ module BinarySearch_Impl0_Len
   use Type
   clone BinarySearch_Impl0_LenLogic as LenLogic0 with type t = t
   use mach.int.Int64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = Type.binarysearch_list t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = usize
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = Type.binarysearch_list t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = Type.binarysearch_list t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = Type.binarysearch_list t
   let rec cfg len (self : Type.binarysearch_list t) : usize
     requires {LenLogic0.len_logic self <= 1000000}
     ensures { UInt64.to_int result = LenLogic0.len_logic self }
@@ -366,11 +366,11 @@ module BinarySearch_BinarySearch
   clone BinarySearch_Impl1_IsSorted as IsSorted0 with function Get0.get = Get0.get
   clone BinarySearch_Impl0_GetDefault as GetDefault0 with type t = uint32, function Get0.get = Get0.get
   clone BinarySearch_Impl0_LenLogic as LenLogic0 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve4 with type self = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = usize
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = Type.binarysearch_list uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve4 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = Type.binarysearch_list uint32
   clone BinarySearch_Impl0_Index_Interface as Index0 with type t = uint32,
   function LenLogic0.len_logic = LenLogic0.len_logic, function Get0.get = Get0.get
   clone BinarySearch_Impl0_Len_Interface as Len0 with type t = uint32,
