+++ creusot/tests/should_succeed/bug/01_resolve_unsoundness.stdout
@@ -28,18 +28,18 @@ module Type
     | CreusotContracts_Std1_Vec_Vec (alloc_vec_vec 't (alloc_alloc_global))
     
 end
-module CreusotContracts_Builtins_Model_Model_ModelTy
+module CreusotContracts_Logic_Model_Model_ModelTy
   type self   
   type modelty   
 end
-module CreusotContracts_Builtins_Model_Model_Model_Interface
+module CreusotContracts_Logic_Model_Model_Model_Interface
   type self   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = self
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
   function model (self : self) : ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Model_Model
+module CreusotContracts_Logic_Model_Model_Model
   type self   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = self
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
   function model (self : self) : ModelTy0.modelty
 end
 module CreusotContracts_Std1_Vec_Impl0_ModelTy
@@ -65,9 +65,9 @@ module CreusotContracts_Std1_Vec_Impl0
   use Type
   clone CreusotContracts_Std1_Vec_Impl0_Model as Model0 with type t = t
   clone CreusotContracts_Std1_Vec_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = Type.creusotcontracts_std1_vec_vec t,
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.creusotcontracts_std1_vec_vec t,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = Type.creusotcontracts_std1_vec_vec t,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.creusotcontracts_std1_vec_vec t,
   type modelty = ModelTy0.modelty
 end
 module CreusotContracts_Std1_Vec_Impl1_New_Interface
@@ -92,48 +92,46 @@ module CreusotContracts_Std1_Vec_Impl1_New
     ensures { Seq.length (Model0.model result) = 0 }
     
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Model_Impl1_ModelTy
+module CreusotContracts_Logic_Model_Impl1_ModelTy
   type t   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = t
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = t
   type modelty  = 
     ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Impl1_Model_Interface
+module CreusotContracts_Logic_Model_Impl1_Model_Interface
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = t
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = t
   function model (self : borrowed t) : ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Impl1_Model
+module CreusotContracts_Logic_Model_Impl1_Model
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = t
-  clone CreusotContracts_Builtins_Model_Model_Model_Interface as Model0 with type self = t,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = t
+  clone CreusotContracts_Logic_Model_Model_Model_Interface as Model0 with type self = t,
   type ModelTy0.modelty = ModelTy0.modelty
   function model (self : borrowed t) : ModelTy0.modelty = 
     Model0.model ( * self)
 end
-module CreusotContracts_Builtins_Model_Impl1
+module CreusotContracts_Logic_Model_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy2 with type self = t
-  clone CreusotContracts_Builtins_Model_Model_Model as Model2 with type self = t,
-  type ModelTy0.modelty = ModelTy2.modelty
-  clone CreusotContracts_Builtins_Model_Impl1_Model as Model0 with type t = t, type ModelTy0.modelty = ModelTy2.modelty,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy2 with type self = t
+  clone CreusotContracts_Logic_Model_Model_Model as Model2 with type self = t, type ModelTy0.modelty = ModelTy2.modelty
+  clone CreusotContracts_Logic_Model_Impl1_Model as Model0 with type t = t, type ModelTy0.modelty = ModelTy2.modelty,
   function Model0.model = Model2.model
-  clone CreusotContracts_Builtins_Model_Impl1_ModelTy as ModelTy0 with type t = t,
-  type ModelTy0.modelty = ModelTy2.modelty
-  clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Model_Impl1_ModelTy as ModelTy0 with type t = t, type ModelTy0.modelty = ModelTy2.modelty
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = borrowed t,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
   type modelty = ModelTy0.modelty
 end
 module CreusotContracts_Std1_Vec_Impl1_Push_Interface
@@ -142,7 +140,7 @@ module CreusotContracts_Std1_Vec_Impl1_Push_Interface
   use prelude.Prelude
   use Type
   clone CreusotContracts_Std1_Vec_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Builtins_Model_Impl1_Model_Interface as Model1 with type t = Type.creusotcontracts_std1_vec_vec t,
+  clone CreusotContracts_Logic_Model_Impl1_Model_Interface as Model1 with type t = Type.creusotcontracts_std1_vec_vec t,
   type ModelTy0.modelty = ModelTy0.modelty
   clone CreusotContracts_Std1_Vec_Impl0_Model_Interface as Model0 with type t = t
   val push (self : borrowed (Type.creusotcontracts_std1_vec_vec t)) (v : t) : ()
@@ -155,7 +153,7 @@ module CreusotContracts_Std1_Vec_Impl1_Push
   use prelude.Prelude
   use Type
   clone CreusotContracts_Std1_Vec_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Builtins_Model_Impl1_Model_Interface as Model1 with type t = Type.creusotcontracts_std1_vec_vec t,
+  clone CreusotContracts_Logic_Model_Impl1_Model_Interface as Model1 with type t = Type.creusotcontracts_std1_vec_vec t,
   type ModelTy0.modelty = ModelTy0.modelty
   clone CreusotContracts_Std1_Vec_Impl0_Model_Interface as Model0 with type t = t
   val push (self : borrowed (Type.creusotcontracts_std1_vec_vec t)) (v : t) : ()
@@ -180,11 +178,11 @@ module C01ResolveUnsoundness_MakeVecOfSize
   use seq.Seq
   use Type
   clone CreusotContracts_Std1_Vec_Impl0_Model as Model0 with type t = bool
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = Type.creusotcontracts_std1_vec_vec bool
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = Type.creusotcontracts_std1_vec_vec bool
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = ()
   clone CreusotContracts_Std1_Vec_Impl0_ModelTy as ModelTy0 with type t = bool
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
-  clone CreusotContracts_Builtins_Model_Impl1_Model as Model1 with type t = Type.creusotcontracts_std1_vec_vec bool,
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Model_Impl1_Model as Model1 with type t = Type.creusotcontracts_std1_vec_vec bool,
   type ModelTy0.modelty = ModelTy0.modelty, function Model0.model = Model0.model
   clone CreusotContracts_Std1_Vec_Impl1_Push_Interface as Push0 with type t = bool,
   function Model0.model = Model0.model, function Model1.model = Model1.model
