+++ creusot/tests/should_succeed/clones/03.stdout
@@ -19,11 +19,11 @@ module C03_Omg
   function omg (x : t) : bool = 
     true
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -37,7 +37,7 @@ end
 module C03_Prog
   type t   
   clone C03_Omg as Omg0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = t
   let rec cfg prog (x : t) : ()
     ensures { Omg0.omg x }
