+++ creusot/tests/should_succeed/clones/04.stdout
@@ -45,11 +45,11 @@ module C04_C
   function c (x : uint32) : bool = 
     x < (50 : uint32) && B0.b x
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -67,7 +67,7 @@ module C04_F
   clone C04_A as A0
   clone C04_B as B0 with function A0.a = A0.a
   clone C04_C as C0 with function B0.b = B0.b
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
   let rec cfg f (x : uint32) : ()
     requires {C0.c x}
