+++ creusot/tests/should_succeed/drop_pair.stdout
@@ -25,53 +25,53 @@ module DropPair_Main
   }
   
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl0_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl0_Resolve_Interface
   type t1   
   type t2   
   predicate resolve (self : (t1, t2))
 end
-module CreusotContracts_Builtins_Resolve_Impl0_Resolve
+module CreusotContracts_Logic_Resolve_Impl0_Resolve
   type t1   
   type t2   
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface as Resolve1 with type self = t2
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface as Resolve0 with type self = t1
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface as Resolve1 with type self = t2
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface as Resolve0 with type self = t1
   predicate resolve (self : (t1, t2)) = 
     Resolve0.resolve (let (a, _) = self in a) && Resolve1.resolve (let (_, a) = self in a)
 end
-module CreusotContracts_Builtins_Resolve_Impl0
+module CreusotContracts_Logic_Resolve_Impl0
   type t1   
   type t2   
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = t2
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = t1
-  clone CreusotContracts_Builtins_Resolve_Impl0_Resolve as Resolve0 with type t1 = t1, type t2 = t2,
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = t2
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = t1
+  clone CreusotContracts_Logic_Resolve_Impl0_Resolve as Resolve0 with type t1 = t1, type t2 = t2,
   predicate Resolve0.resolve = Resolve2.resolve, predicate Resolve1.resolve = Resolve3.resolve
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = (t1, t2),
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = (t1, t2),
   predicate resolve = Resolve0.resolve
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module DropPair_DropPair2_Interface
@@ -84,8 +84,8 @@ module DropPair_DropPair2
   use prelude.Prelude
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Impl0_Resolve as Resolve0 with type t1 = borrowed uint32,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Impl0_Resolve as Resolve0 with type t1 = borrowed uint32,
   type t2 = borrowed uint32, predicate Resolve1.resolve = Resolve1.resolve
   let rec cfg drop_pair2 (x : (borrowed uint32, borrowed uint32)) : () = 
   var _0 : ();
@@ -114,7 +114,7 @@ module DropPair_Drop
   use prelude.Prelude
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = uint32
   let rec cfg drop (x : borrowed uint32) (y : borrowed uint32) : () = 
   var _0 : ();
   var x_1 : borrowed uint32;
@@ -141,7 +141,7 @@ module DropPair_DropPair_Interface
   use mach.int.Int
   use mach.int.UInt32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl0_Resolve_Interface as Resolve0 with type t1 = borrowed uint32,
+  clone CreusotContracts_Logic_Resolve_Impl0_Resolve_Interface as Resolve0 with type t1 = borrowed uint32,
   type t2 = borrowed uint32
   val drop_pair (x : (borrowed uint32, borrowed uint32)) : ()
     ensures {  ^ (let (a, _) = x in a) =  * (let (a, _) = x in a) }
@@ -152,8 +152,8 @@ module DropPair_DropPair
   use mach.int.Int
   use mach.int.UInt32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Impl0_Resolve as Resolve0 with type t1 = borrowed uint32,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Impl0_Resolve as Resolve0 with type t1 = borrowed uint32,
   type t2 = borrowed uint32, predicate Resolve1.resolve = Resolve1.resolve
   let rec cfg drop_pair (x : (borrowed uint32, borrowed uint32)) : ()
     ensures {  ^ (let (a, _) = x in a) =  * (let (a, _) = x in a) }
