+++ creusot/tests/should_succeed/inc_max.stdout
@@ -10,30 +10,30 @@ module Type
   use floating_point.Double
   use prelude.Prelude
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module IncMax_TakeMax_Interface
@@ -48,8 +48,8 @@ module IncMax_TakeMax
   use mach.int.Int
   use mach.int.UInt32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
   let rec cfg take_max (ma : borrowed uint32) (mb : borrowed uint32) : borrowed uint32
     ensures { if  * ma >=  * mb then  * mb =  ^ mb && result = ma else  * ma =  ^ ma && result = mb }
     
@@ -118,10 +118,10 @@ module IncMax_IncMax
   use mach.int.Int
   use mach.int.UInt32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = uint32
   clone IncMax_TakeMax_Interface as TakeMax0
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = uint32
   let rec cfg inc_max (a : uint32) (b : uint32) : ()
     requires {a <= (1000000 : uint32) && b <= (1000000 : uint32)}
