+++ creusot/tests/should_succeed/inc_max_3.stdout
@@ -26,30 +26,30 @@ module IncMax3_Swap
     ensures {  ^ mma =  * mmb &&  ^ mmb =  * mma }
     
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module IncMax3_IncMax3_Interface
@@ -65,11 +65,11 @@ module IncMax3_IncMax3
   use mach.int.Int
   use mach.int.UInt32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve3 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve3 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = ()
   clone IncMax3_Swap_Interface as Swap0
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = borrowed uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = borrowed uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
   let rec cfg inc_max_3 (ma : borrowed uint32) (mb : borrowed uint32) (mc : borrowed uint32) : ()
     requires { * ma <= (1000000 : uint32) &&  * mb <= (1000000 : uint32) &&  * mc <= (1000000 : uint32)}
     ensures {  ^ ma <>  ^ mb &&  ^ mb <>  ^ mc &&  ^ mc <>  ^ ma }
@@ -240,11 +240,11 @@ module IncMax3_TestIncMax3
   use mach.int.Int
   use mach.int.UInt32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = bool
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = bool
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = uint32
   clone IncMax3_IncMax3_Interface as IncMax30
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = uint32
   let rec cfg test_inc_max_3 (a : uint32) (b : uint32) (c : uint32) : ()
     requires {a <= (1000000 : uint32) && b <= (1000000 : uint32) && c <= (1000000 : uint32)}
