+++ creusot/tests/should_succeed/invariant_moves.stdout
@@ -34,30 +34,30 @@ module Type
     
   axiom core_option_option_Some_0_acc : forall a : 't . core_option_option_Some_0 (Core_Option_Option_Some a : core_option_option 't) = a
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module Alloc_Vec_Impl1_Pop_Interface
@@ -90,12 +90,12 @@ module InvariantMoves_TestInvariantMove
   use mach.int.UInt32
   use prelude.Prelude
   use mach.int.Int64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve4 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = Type.alloc_vec_vec uint32 (Type.alloc_alloc_global)
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = Type.core_option_option uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve4 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = Type.alloc_vec_vec uint32 (Type.alloc_alloc_global)
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.core_option_option uint32
   clone Alloc_Vec_Impl1_Pop_Interface as Pop0 with type t = uint32, type a = Type.alloc_alloc_global
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = Type.alloc_vec_vec uint32 (Type.alloc_alloc_global)
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = Type.alloc_vec_vec uint32 (Type.alloc_alloc_global)
   let rec cfg test_invariant_move (x : Type.alloc_vec_vec uint32 (Type.alloc_alloc_global)) : () = 
   var _0 : ();
   var x_1 : Type.alloc_vec_vec uint32 (Type.alloc_alloc_global);
