+++ creusot/tests/should_succeed/list_index_mut.stdout
@@ -32,30 +32,30 @@ module Type
     
   axiom listindexmut_list_List_0_acc : forall a : uint32, b : listindexmut_option (listindexmut_list) . listindexmut_list_List_0 (ListIndexMut_List a b : listindexmut_list) = a
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module ListIndexMut_Len_Interface
@@ -121,11 +121,11 @@ module ListIndexMut_IndexMut
   clone ListIndexMut_Get as Get0
   clone ListIndexMut_Len as Len0
   use mach.int.Int64
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve4 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve3 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve2 with type t = Type.listindexmut_list
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = usize
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = Type.listindexmut_list
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve4 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve3 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve2 with type t = Type.listindexmut_list
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = Type.listindexmut_list
   let rec cfg index_mut (param_l : borrowed (Type.listindexmut_list)) (param_ix : usize) : borrowed uint32
     requires {UInt64.to_int param_ix < Len0.len ( * param_l)}
     ensures { forall i : (int) . 0 <= i && i < Len0.len ( * param_l) && i <> UInt64.to_int param_ix -> Get0.get ( * param_l) i = Get0.get ( ^ param_l) i }
@@ -260,10 +260,10 @@ module ListIndexMut_Write
   use mach.int.UInt32
   clone ListIndexMut_Get as Get0
   clone ListIndexMut_Len as Len0
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve3 with type t = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = usize
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = Type.listindexmut_list
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve3 with type t = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = Type.listindexmut_list
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
   clone ListIndexMut_IndexMut_Interface as IndexMut0 with function Len0.len = Len0.len, function Get0.get = Get0.get
   let rec cfg write (l : borrowed (Type.listindexmut_list)) (ix : usize) (v : uint32) : ()
     requires {UInt64.to_int ix < Len0.len ( * l)}
@@ -320,7 +320,7 @@ end
 module ListIndexMut_Impl0
   use Type
   clone ListIndexMut_Impl0_Resolve as Resolve0
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = Type.listindexmut_list,
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.listindexmut_list,
   predicate resolve = Resolve0.resolve
 end
 module ListIndexMut_Main_Interface
@@ -336,7 +336,7 @@ module ListIndexMut_Main
   clone ListIndexMut_Get as Get0
   clone ListIndexMut_Len as Len0
   clone ListIndexMut_Write_Interface as Write0 with function Len0.len = Len0.len, function Get0.get = Get0.get
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = Type.listindexmut_list
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = Type.listindexmut_list
   let rec cfg main () : () = 
   var _0 : ();
   var l_1 : Type.listindexmut_list;
