+++ creusot/tests/should_succeed/model.stdout
@@ -16,18 +16,18 @@ module Type
     | Model_Pair 't 'u
     
 end
-module CreusotContracts_Builtins_Model_Model_ModelTy
+module CreusotContracts_Logic_Model_Model_ModelTy
   type self   
   type modelty   
 end
-module CreusotContracts_Builtins_Model_Model_Model_Interface
+module CreusotContracts_Logic_Model_Model_Model_Interface
   type self   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = self
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
   function model (self : self) : ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Model_Model
+module CreusotContracts_Logic_Model_Model_Model
   type self   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = self
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
   function model (self : self) : ModelTy0.modelty
 end
 module Model_Impl0_ModelTy
@@ -49,9 +49,9 @@ module Model_Impl0
   use Type
   clone Model_Impl0_Model as Model0
   clone Model_Impl0_ModelTy as ModelTy0
-  clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = Type.model_seven,
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.model_seven,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = Type.model_seven,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.model_seven,
   type modelty = ModelTy0.modelty
 end
 module Model_Seven_Interface
@@ -96,9 +96,9 @@ module Model_Impl1
   use Type
   clone Model_Impl1_Model as Model0 with type t = t, type u = u
   clone Model_Impl1_ModelTy as ModelTy0 with type t = t, type u = u
-  clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = Type.model_pair t u,
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.model_pair t u,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = Type.model_pair t u,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.model_pair t u,
   type modelty = ModelTy0.modelty
 end
 module Model_Pair_Interface
