+++ creusot/tests/should_succeed/modules.stdout
@@ -28,11 +28,11 @@ module Modules_Nested_Further_Another
   }
   
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -48,7 +48,7 @@ end
 module Modules_Nested_Impl0
   use Type
   clone Modules_Nested_Impl0_Resolve as Resolve0
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = Type.modules_nested_nested,
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = Type.modules_nested_nested,
   predicate resolve = Resolve0.resolve
 end
 module Modules_Nested_InnerFunc_Interface
