+++ creusot/tests/should_succeed/projection_toggle.stdout
@@ -10,30 +10,30 @@ module Type
   use floating_point.Double
   use prelude.Prelude
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module ProjectionToggle_ProjToggle_Interface
@@ -46,8 +46,8 @@ end
 module ProjectionToggle_ProjToggle
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = bool
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = bool
   let rec cfg proj_toggle (toggle : bool) (a : borrowed t) (b : borrowed t) : borrowed t
     ensures { if toggle then result = a &&  ^ b =  * b else result = b &&  ^ a =  * a }
     
@@ -110,10 +110,10 @@ module ProjectionToggle_Main
   use mach.int.Int
   use mach.int.Int32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = ()
   clone ProjectionToggle_ProjToggle_Interface as ProjToggle0 with type t = int32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = int32
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = int32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = int32
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = int32
   let rec cfg main () : () = 
   var _0 : ();
   var a_1 : int32;
