+++ creusot/tests/should_succeed/sum.stdout
@@ -25,11 +25,11 @@ module Sum_Main
   }
   
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -43,8 +43,8 @@ end
 module Sum_SumFirstN
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
   let rec cfg sum_first_n (n : uint32) : uint32
     ensures { result = div (n * (n + (1 : uint32))) (2 : uint32) }
