+++ creusot/tests/should_succeed/syntax/02_operators.stdout
@@ -18,11 +18,11 @@ module Type
     
   axiom c02operators_x_X_a_acc : forall a : usize . c02operators_x_X_a (C02Operators_X a : c02operators_x) = a
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -36,7 +36,7 @@ module C02Operators_Division
   use mach.int.Int
   use prelude.Prelude
   use mach.int.UInt64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   let rec cfg division (x : usize) (y : usize) : usize = 
   var _0 : usize;
   var x_1 : usize;
@@ -76,7 +76,7 @@ module C02Operators_Modulus
   use mach.int.Int
   use prelude.Prelude
   use mach.int.UInt64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   let rec cfg modulus (x : usize) (y : usize) : usize = 
   var _0 : usize;
   var x_1 : usize;
@@ -116,7 +116,7 @@ module C02Operators_Multiply
   use mach.int.Int
   use prelude.Prelude
   use mach.int.UInt64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   let rec cfg multiply (x : usize) (y : usize) : usize = 
   var _0 : usize;
   var x_1 : usize;
@@ -150,7 +150,7 @@ module C02Operators_Add
   use mach.int.Int
   use prelude.Prelude
   use mach.int.UInt64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   let rec cfg add (x : usize) (y : usize) : usize = 
   var _0 : usize;
   var x_1 : usize;
@@ -184,7 +184,7 @@ module C02Operators_Sub
   use mach.int.Int
   use prelude.Prelude
   use mach.int.UInt64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   let rec cfg sub (x : usize) (y : usize) : usize = 
   var _0 : usize;
   var x_1 : usize;
@@ -218,7 +218,7 @@ module C02Operators_Expression
   use mach.int.Int
   use prelude.Prelude
   use mach.int.UInt64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = usize
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = usize
   let rec cfg expression (x : usize) (y : usize) (z : usize) : bool = 
   var _0 : bool;
   var x_1 : usize;
@@ -349,7 +349,7 @@ module C02Operators_PrimitiveComparison
   use prelude.Prelude
   use mach.int.UInt64
   use Type
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = Type.c02operators_x
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = Type.c02operators_x
   let rec cfg primitive_comparison (x : Type.c02operators_x) : ()
     ensures { Type.c02operators_x_X_a x <= Type.c02operators_x_X_a x }
     
@@ -374,7 +374,7 @@ module C02Operators_BoolEq_Interface
 end
 module C02Operators_BoolEq
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = bool
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = bool
   let rec cfg bool_eq (a : bool) (b : bool) : bool
     ensures { result = (a = b) }
