+++ creusot/tests/should_succeed/syntax/03_unbounded.stdout
@@ -10,11 +10,11 @@ module Type
   use floating_point.Double
   use prelude.Prelude
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -26,7 +26,7 @@ module C03Unbounded_NoBoundsCheck_Interface
 end
 module C03Unbounded_NoBoundsCheck
   use mach.int.Int
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = int
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = int
   let rec cfg no_bounds_check (x : int) (y : int) : int
     ensures { result = (4294967294 : int) }
