+++ creusot/tests/should_succeed/syntax/04_assoc_prec.stdout
@@ -10,11 +10,11 @@ module Type
   use floating_point.Double
   use prelude.Prelude
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -32,7 +32,7 @@ module C04AssocPrec_RespectPrec
   use mach.int.Int
   use mach.int.UInt32
   use mach.int.Int32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = (uint32, uint32)
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = (uint32, uint32)
   let rec cfg respect_prec (x : (uint32, uint32)) : ()
     ensures { (let (a, _) = x in a) = (let (_, a) = x in a) }
     ensures { div (5 * 3) 2 <> 4 * (40 + 1) }
