+++ creusot/tests/should_succeed/syntax/05_annotations.stdout
@@ -10,11 +10,11 @@ module Type
   use floating_point.Double
   use prelude.Prelude
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -24,8 +24,8 @@ module C05Annotations_Assertion_Interface
 end
 module C05Annotations_Assertion
   type t   
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = ()
   let rec cfg assertion (x : t) : () = 
   var _0 : ();
   var x_1 : t;
