+++ creusot/tests/should_succeed/traits/02.stdout
@@ -24,11 +24,11 @@ module C02_A_IsTrue
     ensures { result = true }
     
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -41,7 +41,7 @@ end
 module C02_Omg
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = t
   clone C02_A_IsTrue_Interface as IsTrue0 with type self = t
   let rec cfg omg (a : t) : bool
     ensures { result = true }
