+++ creusot/tests/should_succeed/traits/03.stdout
@@ -20,11 +20,11 @@ module C03_A_F
   use prelude.Prelude
   val f (self : self) : self
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -38,7 +38,7 @@ module C03_Impl0_F
   use prelude.Prelude
   use mach.int.Int
   use mach.int.Int32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = int32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = int32
   let rec cfg f (self : int32) : int32 = 
   var _0 : int32;
   var self_1 : int32;
@@ -83,7 +83,7 @@ module C03_Impl1_G
   use prelude.Prelude
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
   let rec cfg g (self : uint32) : uint32 = 
   var _0 : uint32;
   var self_1 : uint32;
@@ -124,7 +124,7 @@ end
 module C03_Impl2_H
   type g   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = g
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = g
   let rec cfg h (y : g) : g = 
   var _0 : g;
   var y_1 : g;
