+++ creusot/tests/should_succeed/traits/04.stdout
@@ -40,11 +40,11 @@ module C04_A_Func3
   use prelude.Prelude
   val func3 (self : self) (o : self) : bool
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -59,8 +59,8 @@ module C04_User
   type t   
   use prelude.Prelude
   clone C04_A_Func3_Interface as Func30 with type self = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = bool
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = bool
   clone C04_A_Func2_Interface as Func20 with type self = t
   clone C04_A_Func1_Interface as Func10 with type self = t
   let rec cfg user (a : t) (b : t) : bool
