+++ creusot/tests/should_succeed/traits/08.stdout
@@ -32,11 +32,11 @@ module C08_Tr_Predicate
   use prelude.Prelude
   predicate predicate' (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -48,7 +48,7 @@ end
 module C08_Tr_Program
   type self   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = self
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = self
   let rec cfg program (self : self) : () = 
   var _0 : ();
   var self_1 : self;
@@ -69,7 +69,7 @@ module C08_Test_Interface
 end
 module C08_Test
   type t   
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = t
   let rec cfg test (_1 : t) : () = 
   var _0 : ();
   var _1 : t;
