+++ creusot/tests/should_succeed/traits/11.stdout
@@ -10,11 +10,11 @@ module Type
   use floating_point.Double
   use prelude.Prelude
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -24,7 +24,7 @@ module C11_Test_Interface
 end
 module C11_Test
   type t   
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = t
   let rec cfg test (_1 : t) : () = 
   var _0 : ();
   var _1 : t;
