+++ creusot/tests/should_succeed/traits/12_default_method.stdout
@@ -10,11 +10,11 @@ module Type
   use floating_point.Double
   use prelude.Prelude
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -30,7 +30,7 @@ module C12DefaultMethod_T_Default
   use prelude.Prelude
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = self
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = self
   let rec cfg default (self : self) : uint32 = 
   var _0 : uint32;
   var self_1 : self;
@@ -68,7 +68,7 @@ module C12DefaultMethod_ShouldUseImpl
   clone C12DefaultMethod_T_LogicDefault as LogicDefault0 with type self = uint32
   use prelude.Prelude
   clone C12DefaultMethod_T_Default_Interface as Default0 with type self = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = uint32
   let rec cfg should_use_impl (x : uint32) : ()
     ensures { LogicDefault0.logic_default x }
