+++ creusot/tests/should_succeed/traits/13_assoc_types.stdout
@@ -30,11 +30,11 @@ module C13AssocTypes_Impl0_ModelTy
   type modelty  = 
     ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -48,7 +48,7 @@ module C13AssocTypes_Impl0_Model
   type t   
   use prelude.Prelude
   clone C13AssocTypes_Model_ModelTy as ModelTy0 with type self = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = t
   let rec cfg model (self : t) : ModelTy0.modelty = 
   var _0 : ModelTy0.modelty;
   var self_1 : t;
