+++ creusot/tests/should_succeed/traits/15_impl_interfaces.stdout
@@ -28,11 +28,11 @@ module C15ImplInterfaces_Impl0_A
   type a  = 
     ()
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -46,7 +46,7 @@ end
 module C15ImplInterfaces_Calls
   clone C15ImplInterfaces_Impl0_A as A0
   clone C15ImplInterfaces_X as X0 with type t = (), type A0.a = A0.a
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = ()
   let rec cfg calls (a : ()) : ()
     requires {X0.x a = ()}
