+++ creusot/tests/should_succeed/traits/16_impl_cloning.stdout
@@ -28,18 +28,18 @@ module Type
     | C16ImplCloning_Vec (alloc_vec_vec 't (alloc_alloc_global))
     
 end
-module CreusotContracts_Builtins_Model_Model_ModelTy
+module CreusotContracts_Logic_Model_Model_ModelTy
   type self   
   type modelty   
 end
-module CreusotContracts_Builtins_Model_Model_Model_Interface
+module CreusotContracts_Logic_Model_Model_Model_Interface
   type self   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = self
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
   function model (self : self) : ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Model_Model
+module CreusotContracts_Logic_Model_Model_Model
   type self   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = self
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = self
   function model (self : self) : ModelTy0.modelty
 end
 module C16ImplCloning_Impl0_ModelTy
@@ -65,71 +65,69 @@ module C16ImplCloning_Impl0
   use Type
   clone C16ImplCloning_Impl0_Model as Model0 with type t = t
   clone C16ImplCloning_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = Type.c16implcloning_vec t,
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = Type.c16implcloning_vec t,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = Type.c16implcloning_vec t,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = Type.c16implcloning_vec t,
   type modelty = ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Impl1_ModelTy
+module CreusotContracts_Logic_Model_Impl1_ModelTy
   type t   
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = t
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = t
   type modelty  = 
     ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Impl1_Model_Interface
+module CreusotContracts_Logic_Model_Impl1_Model_Interface
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = t
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = t
   function model (self : borrowed t) : ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Model_Impl1_Model
+module CreusotContracts_Logic_Model_Impl1_Model
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy0 with type self = t
-  clone CreusotContracts_Builtins_Model_Model_Model_Interface as Model0 with type self = t,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy0 with type self = t
+  clone CreusotContracts_Logic_Model_Model_Model_Interface as Model0 with type self = t,
   type ModelTy0.modelty = ModelTy0.modelty
   function model (self : borrowed t) : ModelTy0.modelty = 
     Model0.model ( * self)
 end
-module CreusotContracts_Builtins_Model_Impl1
+module CreusotContracts_Logic_Model_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy2 with type self = t
-  clone CreusotContracts_Builtins_Model_Model_Model as Model2 with type self = t,
-  type ModelTy0.modelty = ModelTy2.modelty
-  clone CreusotContracts_Builtins_Model_Impl1_Model as Model0 with type t = t, type ModelTy0.modelty = ModelTy2.modelty,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy2 with type self = t
+  clone CreusotContracts_Logic_Model_Model_Model as Model2 with type self = t, type ModelTy0.modelty = ModelTy2.modelty
+  clone CreusotContracts_Logic_Model_Impl1_Model as Model0 with type t = t, type ModelTy0.modelty = ModelTy2.modelty,
   function Model0.model = Model2.model
-  clone CreusotContracts_Builtins_Model_Impl1_ModelTy as ModelTy0 with type t = t,
-  type ModelTy0.modelty = ModelTy2.modelty
-  clone CreusotContracts_Builtins_Model_Model_Model as Model1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Model_Impl1_ModelTy as ModelTy0 with type t = t, type ModelTy0.modelty = ModelTy2.modelty
+  clone CreusotContracts_Logic_Model_Model_Model as Model1 with type self = borrowed t,
   type ModelTy0.modelty = ModelTy0.modelty, function model = Model0.model
-  clone CreusotContracts_Builtins_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Model_Model_ModelTy as ModelTy1 with type self = borrowed t,
   type modelty = ModelTy0.modelty
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module C16ImplCloning_Test_Interface
@@ -138,7 +136,7 @@ module C16ImplCloning_Test_Interface
   use Type
   clone C16ImplCloning_Impl0_Model_Interface as Model1 with type t = t
   clone C16ImplCloning_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Builtins_Model_Impl1_Model_Interface as Model0 with type t = Type.c16implcloning_vec t,
+  clone CreusotContracts_Logic_Model_Impl1_Model_Interface as Model0 with type t = Type.c16implcloning_vec t,
   type ModelTy0.modelty = ModelTy0.modelty
   val test (x : borrowed (Type.c16implcloning_vec t)) : ()
     ensures { Model0.model x = Model1.model ( * x) }
@@ -150,9 +148,9 @@ module C16ImplCloning_Test
   use Type
   clone C16ImplCloning_Impl0_Model as Model1 with type t = t
   clone C16ImplCloning_Impl0_ModelTy as ModelTy0 with type t = t
-  clone CreusotContracts_Builtins_Model_Impl1_Model as Model0 with type t = Type.c16implcloning_vec t,
+  clone CreusotContracts_Logic_Model_Impl1_Model as Model0 with type t = Type.c16implcloning_vec t,
   type ModelTy0.modelty = ModelTy0.modelty, function Model0.model = Model1.model
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = Type.c16implcloning_vec t
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = Type.c16implcloning_vec t
   let rec cfg test (x : borrowed (Type.c16implcloning_vec t)) : ()
     ensures { Model0.model x = Model1.model ( * x) }
