+++ creusot/tests/should_succeed/traits/17_impl_refinement.stdout
@@ -30,11 +30,11 @@ module C17ImplRefinement_Tr_MyFunction
     ensures { UInt64.to_int result >= 10 }
     
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -53,7 +53,7 @@ module C17ImplRefinement_Impl0_MyFunction
   use mach.int.Int
   use mach.int.Int32
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = ()
   let rec cfg my_function (self : ()) : usize
     requires {true}
     ensures { UInt64.to_int result >= 15 }
