+++ creusot/tests/should_succeed/unused_in_loop.stdout
@@ -25,11 +25,11 @@ module UnusedInLoop_Main
   }
   
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
@@ -43,9 +43,9 @@ end
 module UnusedInLoop_UnusedInLoop
   use mach.int.Int
   use mach.int.UInt32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = uint32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = bool
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = uint32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = bool
   let rec cfg unused_in_loop (b : bool) : uint32
     ensures { result = (10 : uint32) }
