+++ creusot/tests/should_succeed/while_let.stdout
@@ -14,30 +14,30 @@ module Type
     | WhileLet_Option_None
     
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Resolve_Resolve_Interface
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Resolve_Resolve
+module CreusotContracts_Logic_Resolve_Resolve_Resolve
   type self   
   predicate resolve (self : self)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve_Interface
+module CreusotContracts_Logic_Resolve_Impl1_Resolve_Interface
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t)
 end
-module CreusotContracts_Builtins_Resolve_Impl1_Resolve
+module CreusotContracts_Logic_Resolve_Impl1_Resolve
   type t   
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Resolve_Impl1
+module CreusotContracts_Logic_Resolve_Impl1
   type t   
   use prelude.Prelude
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve0 with type t = t
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve1 with type self = borrowed t,
   predicate resolve = Resolve0.resolve
 end
 module WhileLet_Main_Interface
@@ -49,9 +49,9 @@ module WhileLet_Main
   use Type
   use prelude.Prelude
   use mach.int.Int64
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve2 with type self = ()
-  clone CreusotContracts_Builtins_Resolve_Impl1_Resolve as Resolve1 with type t = Type.whilelet_option int32
-  clone CreusotContracts_Builtins_Resolve_Resolve_Resolve as Resolve0 with type self = Type.whilelet_option int32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve2 with type self = ()
+  clone CreusotContracts_Logic_Resolve_Impl1_Resolve as Resolve1 with type t = Type.whilelet_option int32
+  clone CreusotContracts_Logic_Resolve_Resolve_Resolve as Resolve0 with type self = Type.whilelet_option int32
   let rec cfg main () : () = 
   var _0 : ();
   var a_1 : Type.whilelet_option int32;
